% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hurricane}
\alias{hurricane}
\title{Hurricane names and damage data}
\format{
A data frame with 94 rows and 13 variables:
\describe{
  \item{year}{Year of the hurricane.}
  \item{name}{Hurricane name.}
  \item{masfem}{Name femininity score, from 1 masculine to 11 feminine.}
  \item{min}{Minimum pressure.}
  \item{gender_mf}{Name gender, male or female.}
  \item{category}{Hurricane category.}
  \item{alldeaths}{Total deaths.}
  \item{ndam}{Normalized damage amount.}
  \item{elapsedyrs}{Years elapsed since the hurricane.}
  \item{source}{Data source.}
  \item{masfem_mturk}{Name femininity score from MTurk ratings.}
  \item{wind}{Highest wind speed.}
  \item{ndam15}{Damage amount normalized to 2015.}
}
}
\source{
Original analysis from Jung et al. (2014),
\url{https://www.pnas.org/content/111/24/8782}.
}
\usage{
data(hurricane)
}
\description{
A data set used to study the association between hurricane name femininity
and hurricane fatalities.
}
\keyword{datasets}
