% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pima_package.R
\name{pima-package}
\alias{pima-package}
\title{Post-selection Inference in Multiverse Analysis}
\description{
Runs resampling-based tests jointly to allow for multiple testing in multiverse analysis, providing weak and strong control of the Family-Wise Error Rate and True Discovery Proportions.
As main feature, it applies the PIMA procedure, a general and flexible framework based on the sign-flip score test.
}
\examples{
# COPIARE ESEMPIO DA FUNZIONE PIMA

}
\references{
Girardi, Vesely, Lakens, Altoè, Pastore, Calcagnì, Finos (2024). Post-selection Inference in Multiverse Analysis (PIMA): An Inferential Framework Based on the Sign Flipping Score Test. Psychometrika, doi: 10.1007/s11336-024-09973-6.

Hemerik, Goeman, Finos (2020). Robust testing in generalized linear models by sign flipping score contributions. Journal of the Royal Statistical Society, Series B (Statistical Methodology), doi: 10.1111/rssb.12369.

De Santis, Goeman, Hemerik, Davenport, Finos (2024). Inference in generalized linear models with robustness to misspecified variances. Pre-print, doi: 10.48550/arXiv.2209.13918.
}
\author{
Livio Finos
}
