% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pima_tree.R
\name{pima_tree}
\alias{pima_tree}
\title{Classification Tree with Adjusted p-values}
\usage{
pima_tree(
  res,
  p.values = "adjusted",
  method = "class",
  alpha = 0.05,
  control = rpart::rpart.control(minsplit = 3),
  ...
)
}
\arguments{
\item{res}{A list object with a specific structure containing regression model results. It should contain `mods`, a list of models, and `summary_table`, a data frame with the summary of results, including estimates and p-values.}

\item{p.values}{A char string indicating which type of p-values to use. Options are `"raw"` or `"adjusted"` (default). When `"raw"`, the function uses the p-values from the `summary_table`.}

\item{method}{used in function `rpart::rpart` if `method="class"` it will classify significant p-values at level `alpha`}

\item{alpha}{used only when `method="class"`}

\item{control}{is `rpart.control(minsplit = 3)` by default.}
}
\value{
A plot of the classification tree using `rpart.plot()`.
}
\description{
This function builds a classification tree based on raw or adjusted p-values using the results of a regression model.
}
\examples{
# Example usage (assuming `res` is a pre-computed result object)
# pvalue_tree(res, p.values = "raw")
}
