% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pimads}
\alias{pimads}
\title{Pima Indians diabetes data}
\format{
A data frame with 768 rows and 8 variables:
\describe{
  \item{npreg}{Number of pregnancies.}
  \item{glucose}{Plasma glucose concentration from an oral glucose tolerance test.}
  \item{pressure}{Diastolic blood pressure in mm Hg.}
  \item{skin}{Triceps skin fold thickness in mm.}
  \item{bmi}{Body mass index.}
  \item{pedigree}{Diabetes pedigree function.}
  \item{age}{Age in years.}
  \item{diabetes}{Diabetes diagnosis, with levels \code{neg} and \code{pos}.}
}
}
\usage{
data(pimads)
}
\description{
A data set of diabetes outcomes and clinical predictors for women from the
Pima Indian population.
}
\references{
Smith, J. W., Everhart, J. E., Dickson, W. C., Knowler, W. C.,
and Johannes, R. S. (1988). Using the ADAP learning algorithm to forecast
the onset of diabetes mellitus. In Proceedings of the Symposium on Computer
Applications in Medical Care.
}
\keyword{datasets}
