% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tdp_bound.R
\name{tdp_bound}
\alias{tdp_bound}
\title{True discovery guarantee in multiverse analysis}
\usage{
tdp_bound(obj, by = NULL, comb_funct = "sum", r = 0, conf_level = 0.95, ...)
}
\arguments{
\item{obj}{a \code{jointest} object, e.g., produced by \code{\link{pima}}.}

\item{by}{name of the column of \code{summary_table} used for combining.
If \code{NULL} (default), the TDP is evaluated over all coefficients in all models.}

\item{comb_funct}{combining function: \code{sum} uses the sum of scores as in \code{\link[sumSome]{sumStats}},
while different p-value combinations are defined as in \code{\link[sumSome]{sumPvals}}.}

\item{r}{parameter for Vovk and Wang's p-value combination.}

\item{conf_level}{confidence level.}

\item{...}{further parameters of \code{sumStats} or \code{sumPvals} (truncation parameters and maximum number of iterations of the algorithm).}
}
\value{
Returns a data frame containing a summary for each subset:
number of elements and confidence bound for the TDP.
}
\description{
Computes lower confidence bounds for True Discovery Proportions, when studying one or more parameters within a multiverse of models.
}
\details{
This function builds on the functions \code{sumStats} and \code{sumPvals} of the \strong{sumSome} package, which
combine scores or p-values (by default without truncation).

In the default \code{by = NULL}, the procedure computes a lower confidence bound for the
number of significant effects (non-null coefficients) among all.
Other inputs of the argument \code{by} return analogous bounds, defined by model (\code{"Model"}) or by coefficient (\code{"Coeff"}).
The bounds are simultaneous and remain valid under post-hoc selection.
}
\examples{
# COPIARE PARTE DELL'ESEMPIO DELLA FUNZIONE pima()
}
\references{
Girardi, Vesely, Lakens, Altoè, Pastore, Calcagnì, Finos (2024). Post-selection Inference in Multiverse Analysis (PIMA): An Inferential Framework Based on the Sign Flipping Score Test. Psychometrika, doi: 10.1007/s11336-024-09973-6.

Vesely, Finos, Goeman (2023). Permutation-based true discovery guarantee by sum tests. Journal of the Royal Statistical Society, Series B (Statistical Methodology), doi: 10.1093/jrsssb/qkad019
}
