% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{transf_p}
\alias{transf_p}
\title{Transforming p-values}
\usage{
transf_p(p, method = "raw")
}
\arguments{
\item{p}{a numeric vector with p-values}

\item{method}{the transformation method. Can be one of "raw" (no transformation),
"-log10" or "z". Can also be a custom function.}
}
\value{
the transformed p-values
}
\description{
Transforming p-values
}
\examples{
p <- c(0.01, 0.05, 0.8)
transf_p(p, method = "raw")
transf_p(p, method = "z")
# with custom function
transf_p(p, method = function(x) x/2)
}
