% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global_test.R
\name{global_tests}
\alias{global_tests}
\title{Global tests in multiverse analysis}
\usage{
global_tests(obj, comb_funct = "maxT", by = NULL)
}
\arguments{
\item{obj}{a \code{jointest} object, e.g., produced by \code{\link{pima}}.}

\item{comb_funct}{combining function (see \code{\link[jointest]{combine-methods}} for more datails).}

\item{by}{name of the column of \code{summary_table} used for combining.
If \code{NULL} (default), a single global null hypothesis is tested.}
}
\value{
Returns an object of class \code{jointest}, containing:
\itemize{
\item \code{Tspace}: data frame of test statistics, where columns correspond to tests, and rows to sign-flipping transformations
(the first is the identity).
\item \code{summary_table}: data frame containing a summary for each test:
combined test statistic and p-value.
}
}
\description{
Tests global null hypotheses for weak FWER control, considering one or more parameters within a multiverse of models.
}
\details{
This function builds on the function \code{combine} of the \strong{jointest} package.

In the default \code{by = NULL}, the procedure tests the global null hypothesis
that there are no significant effects (all considered coefficients in all models are null)
against the alternative that there is at least one significant effect
(one non-null coefficient in one model). Other inputs of the argument \code{by}
test analogous global null hypotheses, defined by model (\code{"Model"}) or by coefficient (\code{"Coeff"}).
}
\examples{
# COPIARE PARTE DELL'ESEMPIO DELLA FUNZIONE pima()
}
\references{
Girardi, Vesely, Lakens, Altoè, Pastore, Calcagnì, Finos (2024). Post-selection Inference in Multiverse Analysis (PIMA): An Inferential Framework Based on the Sign Flipping Score Test. Psychometrika, doi: 10.1007/s11336-024-09973-6.
}
