% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\docType{methods}
\name{pima-method}
\alias{pima-method}
\alias{print.pima}
\alias{summary.pima}
\alias{as.pima}
\alias{plot.pima}
\title{Methods for pima objects}
\usage{
\method{print}{pima}(x, n = 4, ...)

\method{summary}{pima}(object, digits = NULL, ...)

as.pima(object, names_obj = NULL, ...)

\method{plot}{pima}(
  x,
  focal = NULL,
  xvar = NULL,
  p.adjusted = TRUE,
  p.transf = "z",
  alpha = 0.05,
  xlab = NULL,
  ylab = NULL,
  regex = FALSE,
  shapes = NULL,
  facet.scales = NULL,
  facet = NULL,
  which.response = NULL,
  ...
)
}
\arguments{
\item{x}{an object of class \code{pima}.}

\item{n}{number of rows to print from the beginning and end of the scenario table.}

\item{...}{additional arguments to be passed}

\item{object}{an object of class \code{pima}.}

\item{digits}{number of digits when rounding. Default to `NULL` thus no rounding.}

\item{names_obj}{a vector of names, its length must be equal to the length of \code{object}}

\item{focal}{a character vector indicating which coefficients to plot. When > 1 coefficient is provided (or NULL) and `xvar` is not provided the `Part. Cor` column is used instead of the `Estimate`.}

\item{xvar}{character indicating the column of the `object$summary_table` to be plotted on the x axis. Default to "Estimate".}

\item{p.adjusted}{logical indicating whether to plot adjusted p-values (\code{TRUE}, default) or raw p-values (\code{FALSE}).}

\item{p.transf}{can be a character vector indicating the transformation to use (see [transf_p()]) or a custom function.}

\item{alpha}{a value between 0 and 1. The plot will mark the p-values smaller than \code{alpha} (0.05 by default). If equal to 0 or 1 nothing will be marked.}

\item{xlab}{character vector indicating the x-axis label. Default to `xvar`}

\item{ylab}{character vector indicating the y-axis label. Default to `p` or `p.adjust.<method>` where method is `object$p.adjust.method`.}

\item{regex}{logical. If \code{TRUE}, the \code{focal} argument is treated as a regular expression to match coefficient names. Default is \code{FALSE}.}

\item{shapes}{a numeric vector of length 2 specifying the ggplot2 shapes to use for non-significant and significant points, respectively. Default is \code{c(4, 19)}.}

\item{facet.scales}{character string indicating if scales should be \code{"fixed"}, \code{"free"}, \code{"free_x"}, or \code{"free_y"}. Default is \code{"free_x"}.}

\item{facet}{a formula for facetting the plot, passed to \code{\link[ggplot2]{facet_grid}}.}

\item{which.response}{a character vector specifying a subset of response variables to be plotted.}
}
\description{
Methods for \code{pima} objects.
The following are methods to extract and manipulate relevant information from
a \code{pima} object.
}
