% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\docType{methods}
\name{pima-method}
\alias{pima-method}
\alias{summary.pima}
\alias{as.pima}
\alias{plot.pima}
\title{Methods for pima objects}
\usage{
\method{summary}{pima}(object, ...)

\method{as}{pima}(object, names_obj = NULL, ...)

\method{plot}{pima}(object, p.values = c("raw", "adjusted"), alpha = 0.05)
}
\arguments{
\item{object}{an object of class \code{pima}.}

\item{...}{additional arguments to be passed}

\item{names_obj}{a vector of names, its length must be equal to the length of \code{object}}

\item{p.values}{use "raw" or "adjusted" (i.e. corrected for selective inference)}

\item{alpha}{a value between 0 and 1. The plot will mark the p-values smaller than \code{alpha} (0.05 by default). If equal to 0 or 1 nothing will be marked.}
}
\description{
Methods for \code{pima} objects. 
The following are methods to extract and manipulate relevant information from
a \code{pima} object.
}
