% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\docType{methods}
\name{pima-method}
\alias{pima-method}
\alias{print.pima}
\alias{summary.pima}
\alias{as.pima}
\alias{plot.pima}
\title{Methods for pima objects}
\usage{
\method{print}{pima}(object, n = 4)

\method{summary}{pima}(object, ...)

\method{as}{pima}(object, names_obj = NULL, ...)

\method{plot}{pima}(
  object,
  focal = NULL,
  xvar = NULL,
  p.adjusted = TRUE,
  p.transf = "-log10",
  alpha = 0.05,
  xlab = NULL,
  ylab = NULL,
  regex = FALSE,
  shapes = NULL,
  ...
)
}
\arguments{
\item{object}{an object of class \code{pima}.}

\item{...}{additional arguments to be passed}

\item{names_obj}{a vector of names, its length must be equal to the length of \code{object}}

\item{focal}{a character vector indicating which coefficients to plot. When > 1 coefficient is provided (or NULL) and `xvar` is not provided the `Part. Cor` column is used instead of the `Estimate`.}

\item{xvar}{character indicating the column of the `object$summary_table` to be plotted on the x axis. Default to "Estimate".}

\item{p.transf}{can be a character vector indicating the transformation to use (see [transf_p()]) or a custom function.}

\item{alpha}{a value between 0 and 1. The plot will mark the p-values smaller than \code{alpha} (0.05 by default). If equal to 0 or 1 nothing will be marked.}

\item{xlab}{character vector indicating the x-axis label. Default to `xvar`}

\item{ylab}{character vector indicating the y-axis label. Default to `p` or `p.adjust.<method>` where method is `object$p.adjust.method`.}

\item{p.adjust}{logical indicating whether plotting raw (`FALSE`) or adjusted p.values (`TRUE`, default).}
}
\description{
Methods for \code{pima} objects.
The following are methods to extract and manipulate relevant information from
a \code{pima} object.
}
