% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pima_package.R
\name{pima-package}
\alias{pima-package}
\title{Post-selection Inference in Multiverse Analysis}
\description{
Runs resampling-based tests jointly to allow for multiple testing in multiverse analysis.
As main feature, it applies the PIMA procedure, a general and flexible framework based on the sign-flip score test
that provides control of the Family-Wise Error Rate (FWER) and confidence bounds for True Discovery Proportions (TDPs).
}
\details{
The starting point of the package is the \code{\link[pima]{pima}} function.
See examples there.
}
\references{
Girardi, Vesely, Lakens, Altoè, Pastore, Calcagnì, Finos (2024). Post-selection Inference in Multiverse Analysis (PIMA): An Inferential Framework Based on the Sign Flipping Score Test. Psychometrika, doi: 10.1007/s11336-024-09973-6.

Hemerik, Goeman, Finos (2020). Robust testing in generalized linear models by sign flipping score contributions. Journal of the Royal Statistical Society, Series B (Statistical Methodology), doi: 10.1111/rssb.12369.

De Santis, Goeman, Hemerik, Davenport, Finos (2025). Inference in generalized linear models with robustness to misspecified variances. Journal of the American Statistical Association, doi: 10.1080/01621459.2025.2491775.
}
\author{
Livio Finos, Paolo Girardi, Filippo Gambarota, Anna Vesely, Giulia Calignano, Massimiliano Pastore, Gianmarco Altoè
}
