% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pima_package.R
\name{pima-package}
\alias{pima-package}
\title{Post-selection Inference in Multiverse Analysis}
\description{
Runs resampling-based tests jointly to allow for multiple testing in multiverse analysis, providing weak and strong control of the Family-Wise Error Rate and True Discovery Proportions.
As main feature, it applies the PIMA procedure, a general and flexible framework based on the sign-flip score test.
}
\examples{
# COPIARE ESEMPIO DA FUNZIONE PIMA

}
\author{
Livio Finos
}
