% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pima.R
\name{pima}
\alias{pima}
\title{Post-selection Inference in Multiverse Analysis}
\usage{
pima(mods, tested_coeffs = NULL, n_flips = 5000, ...)
}
\arguments{
\item{mods}{list of objects that can be evaluated by \code{\link[flipscores]{flipscores}}, usually
model objects produced by \code{glm} or \code{flipscores}.}

\item{tested_coeffs}{tested coefficients. It can be a vector of names (common to all models)
or a list of the same length as \code{mods}, where each element is a vector of names.
If \code{NULL}, all coefficients are tested.}

\item{n_flips}{number of sign flips.}

\item{...}{further parameters of \code{\link[join_flipscores]{join_flipscores}}.}
}
\value{
Returns an object of class \code{pima}, containing:
\itemize{
\item \code{Tspace}: data frame of test statistics, where columns correspond to tested coefficients, and rows to sign-flipping transformations
(the first is the identity).
\item \code{summary_table}: data frame containing a summary for each tested coefficient in each model:
estimate, score, standard error, z value, partial correlation, p-value.
\item \code{mods}: list of model objects computed by \code{flipscores}.
\item \code{tested_coeffs}: tested coefficients, as in input.
}
}
\description{
Applies the PIMA procedure to make selective inference on one or more parameters within a multiverse of models,
derived from different specifications of the data (e.g., pre-processing) and of the model itself (GLMs).
The method jointly computes resampling-based test statistics for all coefficients of interest in all models.
The output allows for multiple testing in multiverse analysis, providing weak and strong control of the Family-Wise Error Rate (FWER)
and confidence bounds for True Discovery Proportions (TDPs).
}
\details{
The procedure builds on the sign-flip score test implemented in the function \code{flipscores} of the \strong{flipscores} package.
For each tested coefficient in each model, it computes a set of resampling-based test statistics and a raw p-value
for the null hypothesis that the coefficient is null against a two-sided alternative.

Post-selection inference is achieved by merging information from the considered models, i.e.,
by combining the resulting test statistics (maximum, mean, sum, etc.):
\itemize{
\item \code{combine} (\strong{jointest}) produces a global p-value for weak FWER control,
possibly by model or by coefficient: "Is there at least one significant effect?"
\item \code{p.adjust} (\strong{jointest}) adds to the output summary multiplicity-adjusted p-values for strong FWER control:
"Which effects in which models are significant?"
\item \code{sumStats} and \code{sumPvals} (\strong{sumSome}) compute a lower confidence bound for the TDP, among all tested effects or within a subset:
"How many of the considered effects are significant?"
}

Further parameters include:
\itemize{
\item \code{score_type}: type of score that is computed (see \code{\link[flipscores]{flipscores}} for more datails).
When the standardized scores are used as test statistics (default \code{score_type = "standardized"}),
the method has exact control of the type I error asymptotically in the sample size
and, for practical purposes, almost exact control for any sample size.
\item \code{statistics}: test statistics computed by the procedure. Currently, \code{t} is the only implemented method.
Different statistics will affect the multivariate inference, but not the univariate.
\item \code{seed}: can be specified to ensure replicability of results.
\item \code{output_models}: \code{TRUE} to return the list of model objects produced by \code{flipscores}.
}
}
\examples{
# Generate data
seed <- 123
set.seed(seed)
n <- 20
D <- data.frame(X1 = rnorm(n), X2 = rnorm(n)^2, Z1 = rnorm(n), Z2 = rnorm(n))
D$Y <- D$X1 + D$Z1 + rnorm(n)

# Multiverse of 4 models (list)
mod1 <- glm(Y ~ X1 + X2 + Z1 + Z2, data = D)
mod2 <- glm(Y ~ X1 + X2 + poly(Z1,2) + Z2, data = D)
mod3 <- glm(Y ~ X1 + X2 + Z1 + poly(Z2,2), data = D)
mod4 <- glm(Y ~ X1 + X2 + poly(Z1,2) + poly(Z2,2), data = D)
mods <- list(mod1 = mod1, mod2 = mod2, mod3 = mod3, mod4 = mod4)

# Test the coefficients of X1 and X2 (raw p-values)
res <- pima(mods, tested_coeffs = c("X1","X2"), seed = seed)
summary(res)

# Global p-values: overall, by model and by coefficient
summary(jointest::combine(res))
summary(jointest::combine(res, by = "Model"))
summary(jointest::combine(res, by = "Coeff"))

# Add multiplicity-adjusted p-values
res <- jointest::p.adjust(res)
summary(res)

# Lower 95\%-confidence bound for the TDP among the 4 coefficients of X1
require(sumSome)
alpha <- 0.05
S <- which(names(res$Tspace) == "X1")
# - using the sum of test statistics
scores <- abs(as.matrix(res$Tspace))
out <- sumSome::sumStats(scores, S, alpha = alpha)
sumSome::tdp(out)
# - using p-value combinations (e.g., Fisher with truncation)
pvalues <- flip::t2p(scores, obs.only = FALSE)
out <- sumSome::sumPvals(pvalues, S, alpha = alpha, type = "fisher")
sumSome::tdp(out)
}
\references{
Girardi, Vesely, Lakens, Altoè, Pastore, Calcagnì, Finos (2024). Post-selection Inference in Multiverse Analysis (PIMA): An Inferential Framework Based on the Sign Flipping Score Test. Psychometrika, doi: 10.1007/s11336-024-09973-6.

Hemerik, Goeman, Finos (2020). Robust testing in generalized linear models by sign flipping score contributions. Journal of the Royal Statistical Society, Series B (Statistical Methodology), doi: 10.1111/rssb.12369.

De Santis, Goeman, Hemerik, Davenport, Finos (2024). Inference in generalized linear models with robustness to misspecified variances. Pre-print, doi: 10.48550/arXiv.2209.13918.
}
