% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pima_tree.R
\name{pima_tree}
\alias{pima_tree}
\title{Classification Tree with Adjusted p-values}
\usage{
pima_tree(res, p.values = "raw", cp = 0.05)
}
\arguments{
\item{res}{A list object with a specific structure containing regression model results. It should contain `mods`, a list of models, and `summary_table`, a data frame with the summary of results, including estimates and p-values.}

\item{p.values}{A char string indicating which type of p-values to use. Options are `"raw"` (default) or `"adjusted"`. When `"raw"`, the function uses the p-values from the `summary_table`. When `"adjusted"`, p-values are adjusted using the `maxT.light` function.}
}
\value{
A plot of the classification tree using `rpart.plot()`.
}
\description{
This function builds a classification tree based on raw or adjusted p-values using the results of a regression model.
}
\examples{
# Example usage (assuming `res` is a pre-computed result object)
# pvalue_tree(res, p.values = "raw")
}
